Add notifications for empty scores and games
