+++ src/index.js
@@ -1,9 +1,8 @@
 import './index.css';
-import {
-  navigation, setHomeGame, addScore, refreshScores,
-} from './modules/events.js';
+import { navigation, addScore, refreshScores } from './modules/events.js';
 import { myGamesTemplate, enterGamePopup } from './modules/html-templates.js';
 import currentGame from './modules/localstorage/current-game.js';
+import { notifyEmptyScores, setHomeGame } from './modules/conditional_rendering.js';
 
 const popupContainer = document.querySelector('.popups-container');
 
@@ -11,6 +10,7 @@ navigation();
 addScore();
 refreshScores();
 myGamesTemplate();
+notifyEmptyScores();
 
 if (currentGame.fetchCurrentGame().gameId) {
   const refreshBtn = document.querySelector('#refresh-button');
