+++ src/modules/load_scores.js
@@ -1,5 +1,3 @@
-import { scoreRefreshSpinnner } from './spinners.js';
-
 const loadScores = async (htmlContainer, scores) => {
   const data = await scores();
   data.forEach((score) => {
@@ -7,7 +5,6 @@ const loadScores = async (htmlContainer, scores) => {
     listItem.textContent = `${score.user}: ${score.score}`;
     htmlContainer.appendChild(listItem);
   });
-  scoreRefreshSpinnner('stop');
 };
 
 export default loadScores;
