+++ src/template.html
@@ -43,6 +43,7 @@
         <form class="add-score-form">
           <input type="text" id="name-input" placeholder="Player name" value="" required/>
           <input type="number" id="score-input" placeholder="Player score" value="" required/>
+          <span class="add-game-first">Please add a game first</span>
           <button type="submit" class="leaderboard-buttons" id="submit-score">Submit</button>
         </form>
       </section>
